🐛 fix(forge-ado): deadline covers the body read; newest build per definition by queueTime
The transport drained nothing before clearing its timer, so a response
whose headers arrived and whose body stalled left the caller's readJson
without a deadline. The body is now buffered inside the same window and
callers receive an equivalent in-memory Response.

latestPerDefinition kept the first row it saw, which was only correct
for a single newest-first page. A branch scan concatenated with its PR
merge-ref scan let an older passing branch build hide a newer failing
merge-ref build from the same pipeline. It now compares queueTime;
undated rows keep first-seen order.

The checks tests move to provider.checks.test.ts to stay inside the
file-size budget; the stub fixture lives in test-helpers.ts.
